Issue #13 Switch back to use tile index as real tile index -e Original/build commit ID: a35b85cfa0c5aaaf163dd939e2ce94321e00276c

+++ menu_config.c
@@ -23,10 +23,10 @@ typedef struct {
     unsigned char menu_item_tiles[8][14];
     unsigned char menu_item_palette[8][14];
 
-    unsigned short current_item_x;
-    unsigned short current_item_y;
-    unsigned short max_items_x;
-    unsigned short max_items_y;
+    unsigned int current_item_x;
+    unsigned int current_item_y;
+    unsigned int max_items_x;
+    unsigned int max_items_y;
     
     unsigned int tile_offset;
 } menu_config_t;
